+++ compositor_pipeline/src/pipeline/output.rs
@@ -1,34 +1,197 @@
-use compositor_render::OutputId;
-use crossbeam_channel::Receiver;
+use compositor_render::{Frame, OutputFrameFormat, OutputId, Resolution};
+use crossbeam_channel::{bounded, Receiver, Sender};
 
-use crate::error::OutputInitError;
+use crate::{audio_mixer::OutputSamples, error::RegisterOutputError, queue::PipelineEvent};
 
 use self::rtp::{RtpSender, RtpSenderOptions};
 
-use super::{structs::EncoderOutputEvent, Port};
+use super::{
+    encoder::{AudioEncoderOptions, Encoder, EncoderOptions, VideoEncoderOptions},
+    types::EncoderOutputEvent,
+    PipelineCtx, Port, RawDataReceiver,
+};
 
 pub mod rtp;
 
-#[derive(Debug)]
-pub enum Output {
-    Rtp(RtpSender),
+/// Options to configure public outputs that can be constructed via REST API
+#[derive(Debug, Clone)]
+pub struct OutputOptions {
+    pub output_protocol: OutputProtocolOptions,
+    pub video: Option<VideoEncoderOptions>,
+    pub audio: Option<AudioEncoderOptions>,
 }
 
 #[derive(Debug, Clone)]
-pub enum OutputOptions {
+pub enum OutputProtocolOptions {
     Rtp(RtpSenderOptions),
 }
 
-impl Output {
-    pub fn new(
+/// Options to configure output that sends h264 and opus audio via channel
+#[derive(Debug, Clone)]
+pub struct EncodedDataOutputOptions {
+    pub video: Option<VideoEncoderOptions>,
+    pub audio: Option<AudioEncoderOptions>,
+}
+
+/// Options to configure output that sends raw PCM audio + wgpu textures via channel
+#[derive(Debug, Clone)]
+pub struct RawDataOutputOptions {
+    pub video: Option<RawVideoOptions>,
+    pub audio: Option<RawAudioOptions>,
+}
+
+/// Options to configure audio output that returns raw video via channel.
+///
+/// TODO: add option, for now it implies RGBA wgpu::Texture
+#[derive(Debug, Clone)]
+pub struct RawVideoOptions {
+    resolution: Resolution,
+}
+
+/// Options to configure audio output that returns raw audio via channel.
+///
+/// TODO: add option, for now it implies 16-bit stereo
+#[derive(Debug, Clone)]
+pub struct RawAudioOptions;
+
+pub enum Output {
+    Rtp {
+        sender: RtpSender,
+        encoder: Encoder,
+    },
+    EncodedData {
+        encoder: Encoder,
+    },
+    RawData {
+        resolution: Option<Resolution>,
+        video: Option<Sender<PipelineEvent<Frame>>>,
+        audio: Option<Sender<PipelineEvent<OutputSamples>>>,
+    },
+}
+
+pub(super) trait OutputOptionsExt<NewOutputResult> {
+    fn new_output(
+        &self,
         output_id: &OutputId,
-        options: OutputOptions,
-        packets: Receiver<EncoderOutputEvent>,
-    ) -> Result<(Self, Option<Port>), OutputInitError> {
-        match options {
-            OutputOptions::Rtp(options) => {
-                let (sender, port) = rtp::RtpSender::new(output_id, options, packets)?;
-                Ok((Self::Rtp(sender), port))
+        ctx: &PipelineCtx,
+    ) -> Result<(Output, NewOutputResult), RegisterOutputError>;
+}
+
+impl OutputOptionsExt<Option<Port>> for OutputOptions {
+    fn new_output(
+        &self,
+        output_id: &OutputId,
+        ctx: &PipelineCtx,
+    ) -> Result<(Output, Option<Port>), RegisterOutputError> {
+        let encoder_opts = EncoderOptions {
+            video: self.video.clone(),
+            audio: self.audio.clone(),
+        };
+
+        let (encoder, packets) = Encoder::new(output_id, encoder_opts, ctx.output_sample_rate)
+            .map_err(|e| RegisterOutputError::EncoderError(output_id.clone(), e))?;
+
+        match &self.output_protocol {
+            OutputProtocolOptions::Rtp(rtp_options) => {
+                let (sender, port) =
+                    rtp::RtpSender::new(output_id, rtp_options.clone(), packets)
+                        .map_err(|e| RegisterOutputError::OutputError(output_id.clone(), e))?;
+
+                Ok((Output::Rtp { sender, encoder }, port))
+            }
+        }
+    }
+}
+
+impl OutputOptionsExt<Receiver<EncoderOutputEvent>> for EncodedDataOutputOptions {
+    fn new_output(
+        &self,
+        output_id: &OutputId,
+        ctx: &PipelineCtx,
+    ) -> Result<(Output, Receiver<EncoderOutputEvent>), RegisterOutputError> {
+        let encoder_opts = EncoderOptions {
+            video: self.video.clone(),
+            audio: self.audio.clone(),
+        };
+
+        let (encoder, packets) = Encoder::new(output_id, encoder_opts, ctx.output_sample_rate)
+            .map_err(|e| RegisterOutputError::EncoderError(output_id.clone(), e))?;
+
+        Ok((Output::EncodedData { encoder }, packets))
+    }
+}
+
+impl OutputOptionsExt<RawDataReceiver> for RawDataOutputOptions {
+    fn new_output(
+        &self,
+        _output_id: &OutputId,
+        _ctx: &PipelineCtx,
+    ) -> Result<(Output, RawDataReceiver), RegisterOutputError> {
+        let (video_sender, video_receiver, resolution) = match &self.video {
+            Some(opts) => {
+                let (sender, receiver) = bounded(100);
+                (Some(sender), Some(receiver), Some(opts.resolution))
+            }
+            None => (None, None, None),
+        };
+        let (audio_sender, audio_receiver) = match self.audio {
+            Some(_) => {
+                let (sender, receiver) = bounded(100);
+                (Some(sender), Some(receiver))
+            }
+            None => (None, None),
+        };
+        Ok((
+            Output::RawData {
+                resolution,
+                video: video_sender,
+                audio: audio_sender,
+            },
+            RawDataReceiver {
+                video: video_receiver,
+                audio: audio_receiver,
+            },
+        ))
+    }
+}
+
+impl Output {
+    pub fn frame_sender(&self) -> Option<&Sender<PipelineEvent<Frame>>> {
+        match &self {
+            Output::Rtp { encoder, .. } => encoder.frame_sender(),
+            Output::EncodedData { encoder } => encoder.frame_sender(),
+            Output::RawData { video, .. } => video.as_ref(),
+        }
+    }
+
+    pub fn samples_batch_sender(&self) -> Option<&Sender<PipelineEvent<OutputSamples>>> {
+        match &self {
+            Output::Rtp { encoder, .. } => encoder.samples_batch_sender(),
+            Output::EncodedData { encoder } => encoder.samples_batch_sender(),
+            Output::RawData { audio, .. } => audio.as_ref(),
+        }
+    }
+
+    pub fn resolution(&self) -> Option<Resolution> {
+        match &self {
+            Output::Rtp { encoder, .. } => encoder.video.as_ref().map(|v| v.resolution()),
+            Output::EncodedData { encoder } => encoder.video.as_ref().map(|v| v.resolution()),
+            Output::RawData { resolution, .. } => *resolution,
+        }
+    }
+
+    pub(super) fn output_frame_format(&self) -> Option<OutputFrameFormat> {
+        match &self {
+            Output::Rtp { encoder, .. } => encoder
+                .video
+                .as_ref()
+                .map(|_| OutputFrameFormat::PlanarYuv420Bytes),
+            Output::EncodedData { encoder } => encoder
+                .video
+                .as_ref()
+                .map(|_| OutputFrameFormat::PlanarYuv420Bytes),
+            Output::RawData { video, .. } => {
+                video.as_ref().map(|_| OutputFrameFormat::RgbaWgpuTexture)
             }
         }
     }