+++ compositor_render/src/state/render_graph.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 use crate::scene::{self, OutputNode};
 use crate::wgpu::texture::{InputTexture, NodeTexture, OutputTexture};
 use crate::{error::UpdateSceneError, wgpu::WgpuErrorScope};
-use crate::{InputId, OutputId};
+use crate::{InputId, OutputFrameFormat, OutputId};
 
 use super::{node::RenderNode, RenderCtx};
 
@@ -15,6 +15,7 @@ pub(super) struct RenderGraph {
 pub(super) struct OutputRenderTree {
     pub(super) root: RenderNode,
     pub(super) output_texture: OutputTexture,
+    pub(super) output_format: OutputFrameFormat,
 }
 
 impl RenderGraph {
@@ -42,6 +43,7 @@ impl RenderGraph {
         &mut self,
         ctx: &RenderCtx,
         output: OutputNode,
+        output_format: OutputFrameFormat,
     ) -> Result<(), UpdateSceneError> {
         // TODO: If we want nodes to be stateful we could try reusing nodes instead
         //       of recreating them on every scene update
@@ -50,6 +52,7 @@ impl RenderGraph {
         let output_tree = OutputRenderTree {
             root: Self::create_node(ctx, output.node)?,
             output_texture: OutputTexture::new(ctx.wgpu_ctx, output.resolution),
+            output_format,
         };
 
         scope.pop(&ctx.wgpu_ctx.device)?;