+++ compositor_render/src/state/render_loop.rs
@@ -1,12 +1,11 @@
 use std::{collections::HashMap, time::Duration};
 
-use log::error;
+use tracing::{error, warn};
 
 use crate::{
-    scene::RGBColor,
     state::{node::RenderNode, render_graph::RenderGraph, RenderCtx},
-    wgpu::texture::{InputTexture, NodeTexture},
-    Frame, FrameSet, InputId, OutputId,
+    wgpu::texture::{InputTexture, NodeTexture, PlanarYuvPendingDownload},
+    Frame, FrameData, FrameSet, InputId, OutputFrameFormat, OutputId, Resolution,
 };
 
 pub(super) fn populate_inputs(
@@ -34,66 +33,101 @@ pub(super) fn populate_inputs(
     }
 }
 
+enum PartialOutputFrame<'a, F>
+where
+    F: FnOnce() -> Result<bytes::Bytes, wgpu::BufferAsyncError> + 'a,
+{
+    PendingYuvDownload {
+        output_id: OutputId,
+        pending_download: PlanarYuvPendingDownload<'a, F, wgpu::BufferAsyncError>,
+        resolution: Resolution,
+    },
+    CompleteFrame {
+        output_id: OutputId,
+        frame: Frame,
+    },
+}
+
 pub(super) fn read_outputs(
     ctx: &RenderCtx,
     scene: &mut RenderGraph,
     pts: Duration,
 ) -> HashMap<OutputId, Frame> {
-    let mut pending_downloads = Vec::with_capacity(scene.outputs.len());
+    let mut partial_textures = Vec::with_capacity(scene.outputs.len());
     for (output_id, output) in &scene.outputs {
         match output.root.output_texture(&scene.inputs).state() {
-            Some(node) => {
-                ctx.wgpu_ctx.format.convert_rgba_to_yuv(
-                    ctx.wgpu_ctx,
-                    (node.rgba_texture(), node.bind_group()),
-                    output.output_texture.yuv_textures(),
-                );
-            }
+            Some(node) => match output.output_format {
+                OutputFrameFormat::PlanarYuv420Bytes => {
+                    ctx.wgpu_ctx.format.convert_rgba_to_yuv(
+                        ctx.wgpu_ctx,
+                        (node.rgba_texture(), node.bind_group()),
+                        output.output_texture.yuv_textures(),
+                    );
+                    let pending_download = output.output_texture.start_download(ctx.wgpu_ctx);
+                    partial_textures.push(PartialOutputFrame::PendingYuvDownload {
+                        output_id: output_id.clone(),
+                        pending_download,
+                        resolution: output.output_texture.resolution().to_owned(),
+                    });
+                }
+                OutputFrameFormat::RgbaWgpuTexture => {
+                    let texture = node
+                        .rgba_texture()
+                        .texture()
+                        .copy_to_wgpu_texture(ctx.wgpu_ctx);
+                    let size = texture.size();
+                    let frame = Frame {
+                        data: FrameData::Rgba8UnormWgpuTexture(texture.into()),
+                        resolution: Resolution {
+                            width: size.width as usize,
+                            height: size.height as usize,
+                        },
+                        pts,
+                    };
+                    partial_textures.push(PartialOutputFrame::CompleteFrame {
+                        output_id: output_id.clone(),
+                        frame,
+                    })
+                }
+            },
             None => {
-                let (y, u, v) = RGBColor::BLACK.to_yuv();
-                ctx.wgpu_ctx.utils.fill_r8_with_value(
-                    ctx.wgpu_ctx,
-                    output.output_texture.yuv_textures().plane(0),
-                    y,
-                );
-                ctx.wgpu_ctx.utils.fill_r8_with_value(
-                    ctx.wgpu_ctx,
-                    output.output_texture.yuv_textures().plane(1),
-                    u,
-                );
-                ctx.wgpu_ctx.utils.fill_r8_with_value(
-                    ctx.wgpu_ctx,
-                    output.output_texture.yuv_textures().plane(2),
-                    v,
-                );
-            }
-        };
-        let yuv_pending = output.output_texture.start_download(ctx.wgpu_ctx);
-        pending_downloads.push((
-            output_id,
-            yuv_pending,
-            output.output_texture.resolution().to_owned(),
-        ));
-    }
-    ctx.wgpu_ctx.device.poll(wgpu::MaintainBase::Wait);
-
-    let mut result = HashMap::new();
-    for (output_id, yuv_pending, resolution) in pending_downloads {
-        let yuv_data = match yuv_pending.wait() {
-            Ok(data) => data,
-            Err(err) => {
-                error!("Failed to download frame: {}", err);
+                warn!(?output_id, "Missing output texture");
+                // registering output is not possible without a scene so this
+                // should never happen
                 continue;
             }
         };
-        result.insert(
-            output_id.clone(),
-            Frame {
-                data: yuv_data,
+    }
+
+    ctx.wgpu_ctx.device.poll(wgpu::MaintainBase::Wait);
+
+    let mut result = HashMap::new();
+    for partial in partial_textures {
+        match partial {
+            PartialOutputFrame::PendingYuvDownload {
+                output_id,
+                pending_download,
                 resolution,
-                pts,
-            },
-        );
+            } => {
+                let data = match pending_download.wait() {
+                    Ok(data) => data,
+                    Err(err) => {
+                        error!("Failed to download frame: {}", err);
+                        continue;
+                    }
+                };
+                let frame = Frame {
+                    data,
+                    resolution,
+                    pts,
+                };
+                result.insert(output_id.clone(), frame);
+            }
+
+            PartialOutputFrame::CompleteFrame { output_id, frame } => {
+                result.insert(output_id, frame);
+            }
+        }
     }
     result
 }